Implement crud for user

+++ Infrastructure/Persistence/Repositories/UserRepository.cs
@@ -13,9 +13,9 @@ public class UserRepository : IUserRepository
     public async Task<IEnumerable<User>> GetAllAsync(CancellationToken cancellationToken = default) =>
         await _dbContext.Users.ToListAsync(cancellationToken);
 
-    public async Task<User> GetByIdAsync(Guid UserId, CancellationToken cancellationToken = default)
+    public async Task<User> GetByIdAsync(Guid userId, CancellationToken cancellationToken = default)
     {
-        var user = await _dbContext.Users.FirstOrDefaultAsync(x => x.Id == UserId, cancellationToken);
+        var user = await _dbContext.Users.FirstOrDefaultAsync(x => x.Id == userId, cancellationToken);
         return user ?? default;
     }
 
@@ -26,9 +26,19 @@ public class UserRepository : IUserRepository
         return entityEntry.Entity.Id;
     }
 
-    public async Task RemoveAsync(User user, CancellationToken cancellationToken = default)
+    public async Task RemoveAsync(Guid userId, CancellationToken cancellationToken = default)
     {
-        _dbContext.Users.Remove(user);
+        var user = await _dbContext.Users.FindAsync(new object[] { userId }, cancellationToken);
+        if (user != null)
+        {
+            _dbContext.Users.Remove(user);
+            await _dbContext.SaveChangesAsync(cancellationToken);
+        }
+    }
+
+    public async Task UpdateAsync(User user, CancellationToken cancellationToken = default)
+    {
+        _dbContext.Users.Update(user);
         await _dbContext.SaveChangesAsync(cancellationToken);
     }
 }
